+++ lang/da/layouts.php
@@ -5,8 +5,12 @@ return [
     'modal.link_text' => 'Feedback',
     'modal.success_message' => 'Tak for din feedback!',
 
+    'form.name' => 'Navn',
+    'form.email' => 'E-mail',
+    'form.phone' => 'Telefon',
+    'form.note' => 'Besked',
     'form.message' => 'Beskriv dit problem',
-    'form.email' => 'Efterlad din email til kontakt (valgfrit)',
+    'form.optional' => 'valgfrit',
     'form.submit' => 'Send feedback',
 
     'email.subject' => 'Feedback :app_www',