+++ resources/views/emails/feedback-html.blade.php
@@ -1,12 +1,31 @@
 @extends('ig-common::layouts.email-html')
 
 @section('content')
+@if(isset($feedback['name']))
 <p>
-    <strong>@lang('ig-feedback::layouts.form.message'):</strong><br/>
-    {{ $feedback['message'] ?? __('ig-feedback::layouts.email.no_message') }}
+    <strong>@lang('ig-feedback::layouts.form.name'):</strong><br/>
+    {{ $feedback['name'] }}
 </p>
+@endif
+
+@if(isset($feedback['email']))
 <p>
     <strong>@lang('ig-feedback::layouts.form.email'):</strong><br/>
-    {{ $feedback['email'] ?? __('ig-feedback::layouts.email.anonymous') }}
+    {{ $feedback['email'] }}
 </p>
+@endif
+
+@if(isset($feedback['phone']))
+<p>
+    <strong>@lang('ig-feedback::layouts.form.phone'):</strong><br/>
+    {{ $feedback['phone'] }}
+</p>
+@endif
+
+@if(isset($feedback['note']))
+<p>
+    <strong>@lang('ig-feedback::layouts.form.note'):</strong><br/>
+    {{ $feedback['note'] }}
+</p>
+@endif
 @endsection