+++ routes/web.php
@@ -1,6 +1,0 @@
-<?php
-
-use Illuminate\Support\Facades\Route;
-use InternetGuru\LaravelFeedback\Http\Controllers\FeedbackController;
-
-Route::post('/feedback/send', [FeedbackController::class, 'send'])->name('feedback.send');